Implementación CRUD Empleador ok

+++ src/main/java/cl/awakelab/SprintM6/service/IEmpleadorService.java
@@ -5,9 +5,10 @@ import cl.awakelab.SprintM6.entity.Empleador;
 import java.util.List;
 
 public interface IEmpleadorService {
+
     List<Empleador> listarEmpleadores();
     Empleador crearEmpleador (Empleador empleador);
     Empleador buscarEmpleadorPorId (int idEmpleador);
-    Empleador actualizarEmpleador (Empleador empleador, long idEmpleador);
+    Empleador actualizarEmpleador (Empleador empleador, int idEmpleador);
     public void eliminarEmpleador (int idEmpleador);
 }
